testing adding support for unique urls between urls
right not duplicate urls in file (each run by a separate worker) will be run twice.
We can improve upon this by doing the extraction of URLs first, and only adding a url
to a new checker list given that it has not been seen before. This is supported by
adding an optional urls variable to the UrlCheckResult, and only if not defined do we
parse the file. This allows us to parse on a first pass, update with unique urls,
and then hand those urls directly to the worker instance to check

Signed-off-by: vsoch <vsoch@users.noreply.github.com>