bug: update selenium to use newer interface
The current failures are a result of an update to selenium,
so the instantiation of our driver fails, returns as None,
and then all the requests are done with only requests. As
the web matures (and sites do not want scraping) it is less
likely this approach will work - we need the driver. This
change will update the selenium UI to ensure the driver works
and restore functionality. I will follow up with any tweaks
needed for the CI (working locally for me).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>